refactored test modules
